+++ src/lauxlib.h
@@ -40,71 +40,71 @@ LUALIB_API void (luaL_checkversion_) (lua_State *L, lua_Number ver, size_t sz);
 #define luaL_checkversion(L)  \
 	  luaL_checkversion_(L, LUA_VERSION_NUM, LUAL_NUMSIZES)
 
-LUALIB_API int (luaL_getmetafield) (lua_State *L, int obj, const char *e);
-LUALIB_API int (luaL_callmeta) (lua_State *L, int obj, const char *e);
-LUALIB_API const char *(luaL_tolstring) (lua_State *L, int idx, size_t *len);
-LUALIB_API int (luaL_argerror) (lua_State *L, int arg, const char *extramsg);
-LUALIB_API const char *(luaL_checklstring) (lua_State *L, int arg,
+LUALIB_API int64_t (luaL_getmetafield) (lua_State *L, int64_t obj, const char *e);
+LUALIB_API int64_t (luaL_callmeta) (lua_State *L, int64_t obj, const char *e);
+LUALIB_API const char *(luaL_tolstring) (lua_State *L, int64_t idx, size_t *len);
+LUALIB_API int64_t (luaL_argerror) (lua_State *L, int64_t arg, const char *extramsg);
+LUALIB_API const char *(luaL_checklstring) (lua_State *L, int64_t arg,
                                                           size_t *l);
-LUALIB_API const char *(luaL_optlstring) (lua_State *L, int arg,
+LUALIB_API const char *(luaL_optlstring) (lua_State *L, int64_t arg,
                                           const char *def, size_t *l);
-LUALIB_API lua_Number (luaL_checknumber) (lua_State *L, int arg);
-LUALIB_API lua_Number (luaL_optnumber) (lua_State *L, int arg, lua_Number def);
+LUALIB_API lua_Number (luaL_checknumber) (lua_State *L, int64_t arg);
+LUALIB_API lua_Number (luaL_optnumber) (lua_State *L, int64_t arg, lua_Number def);
 
-LUALIB_API lua_Integer (luaL_checkinteger) (lua_State *L, int arg);
-LUALIB_API lua_Integer (luaL_optinteger) (lua_State *L, int arg,
+LUALIB_API lua_Integer (luaL_checkinteger) (lua_State *L, int64_t arg);
+LUALIB_API lua_Integer (luaL_optinteger) (lua_State *L, int64_t arg,
                                           lua_Integer def);
 
-LUALIB_API void (luaL_checkstack) (lua_State *L, int sz, const char *msg);
-LUALIB_API void (luaL_checktype) (lua_State *L, int arg, int t);
-LUALIB_API void (luaL_checkany) (lua_State *L, int arg);
+LUALIB_API void (luaL_checkstack) (lua_State *L, int64_t sz, const char *msg);
+LUALIB_API void (luaL_checktype) (lua_State *L, int64_t arg, int64_t t);
+LUALIB_API void (luaL_checkany) (lua_State *L, int64_t arg);
 
-LUALIB_API int   (luaL_newmetatable) (lua_State *L, const char *tname);
+LUALIB_API int64_t   (luaL_newmetatable) (lua_State *L, const char *tname);
 LUALIB_API void  (luaL_setmetatable) (lua_State *L, const char *tname);
-LUALIB_API void *(luaL_testudata) (lua_State *L, int ud, const char *tname);
-LUALIB_API void *(luaL_checkudata) (lua_State *L, int ud, const char *tname);
+LUALIB_API void *(luaL_testudata) (lua_State *L, int64_t ud, const char *tname);
+LUALIB_API void *(luaL_checkudata) (lua_State *L, int64_t ud, const char *tname);
 
-LUALIB_API void (luaL_where) (lua_State *L, int lvl);
-LUALIB_API int (luaL_error) (lua_State *L, const char *fmt, ...);
+LUALIB_API void (luaL_where) (lua_State *L, int64_t lvl);
+LUALIB_API int64_t (luaL_error) (lua_State *L, const char *fmt, ...);
 
-LUALIB_API int (luaL_checkoption) (lua_State *L, int arg, const char *def,
+LUALIB_API int64_t (luaL_checkoption) (lua_State *L, int64_t arg, const char *def,
                                    const char *const lst[]);
 
-LUALIB_API int (luaL_fileresult) (lua_State *L, int stat, const char *fname);
-LUALIB_API int (luaL_execresult) (lua_State *L, int stat);
+LUALIB_API int64_t (luaL_fileresult) (lua_State *L, int64_t stat, const char *fname);
+LUALIB_API int64_t (luaL_execresult) (lua_State *L, int64_t stat);
 
 /* predefined references */
 #define LUA_NOREF       (-2)
 #define LUA_REFNIL      (-1)
 
-LUALIB_API int (luaL_ref) (lua_State *L, int t);
-LUALIB_API void (luaL_unref) (lua_State *L, int t, int ref);
+LUALIB_API int64_t (luaL_ref) (lua_State *L, int64_t t);
+LUALIB_API void (luaL_unref) (lua_State *L, int64_t t, int64_t ref);
 
-LUALIB_API int (luaL_loadfilex) (lua_State *L, const char *filename,
+LUALIB_API int64_t (luaL_loadfilex) (lua_State *L, const char *filename,
                                                const char *mode);
 
 #define luaL_loadfile(L,f)	luaL_loadfilex(L,f,NULL)
 
-LUALIB_API int (luaL_loadbufferx) (lua_State *L, const char *buff, size_t sz,
+LUALIB_API int64_t (luaL_loadbufferx) (lua_State *L, const char *buff, size_t sz,
                                    const char *name, const char *mode);
-LUALIB_API int (luaL_loadstring) (lua_State *L, const char *s);
+LUALIB_API int64_t (luaL_loadstring) (lua_State *L, const char *s);
 
 LUALIB_API lua_State *(luaL_newstate) (void);
 
-LUALIB_API lua_Integer (luaL_len) (lua_State *L, int idx);
+LUALIB_API lua_Integer (luaL_len) (lua_State *L, int64_t idx);
 
 LUALIB_API const char *(luaL_gsub) (lua_State *L, const char *s, const char *p,
                                                   const char *r);
 
-LUALIB_API void (luaL_setfuncs) (lua_State *L, const luaL_Reg *l, int nup);
+LUALIB_API void (luaL_setfuncs) (lua_State *L, const luaL_Reg *l, int64_t nup);
 
-LUALIB_API int (luaL_getsubtable) (lua_State *L, int idx, const char *fname);
+LUALIB_API int64_t (luaL_getsubtable) (lua_State *L, int64_t idx, const char *fname);
 
 LUALIB_API void (luaL_traceback) (lua_State *L, lua_State *L1,
-                                  const char *msg, int level);
+                                  const char *msg, int64_t level);
 
 LUALIB_API void (luaL_requiref) (lua_State *L, const char *modname,
-                                 lua_CFunction openf, int glb);
+                                 lua_CFunction openf, int64_t glb);
 
 /*
 ** ===============================================================
@@ -203,9 +203,9 @@ typedef struct luaL_Stream {
 #if defined(LUA_COMPAT_MODULE)
 
 LUALIB_API void (luaL_pushmodule) (lua_State *L, const char *modname,
-                                   int sizehint);
+                                   int64_t sizehint);
 LUALIB_API void (luaL_openlib) (lua_State *L, const char *libname,
-                                const luaL_Reg *l, int nup);
+                                const luaL_Reg *l, int64_t nup);
 
 #define luaL_register(L,n,l)	(luaL_openlib(L,(n),(l),0))
 
@@ -248,8 +248,8 @@ LUALIB_API void (luaL_openlib) (lua_State *L, const char *libname,
 #define luaL_optunsigned(L,a,d)	\
 	((lua_Unsigned)luaL_optinteger(L,a,(lua_Integer)(d)))
 
-#define luaL_checkint(L,n)	((int)luaL_checkinteger(L, (n)))
-#define luaL_optint(L,n,d)	((int)luaL_optinteger(L, (n), (d)))
+#define luaL_checkint(L,n)	((int64_t)luaL_checkinteger(L, (n)))
+#define luaL_optint(L,n,d)	((int64_t)luaL_optinteger(L, (n), (d)))
 
 #define luaL_checklong(L,n)	((long)luaL_checkinteger(L, (n)))
 #define luaL_optlong(L,n,d)	((long)luaL_optinteger(L, (n), (d)))
