+++ src/lcorolib.c
@@ -25,8 +25,8 @@ static lua_State *getco (lua_State *L) {
 }
 
 
-static int auxresume (lua_State *L, lua_State *co, int narg) {
-  int status;
+static int64_t auxresume (lua_State *L, lua_State *co, int64_t narg) {
+  int64_t status;
   if (!lua_checkstack(co, narg)) {
     lua_pushliteral(L, "too many arguments to resume");
     return -1;  /* error flag */
@@ -38,7 +38,7 @@ static int auxresume (lua_State *L, lua_State *co, int narg) {
   lua_xmove(L, co, narg);
   status = lua_resume(co, L, narg);
   if (status == LUA_OK || status == LUA_YIELD) {
-    int nres = lua_gettop(co);
+    int64_t nres = lua_gettop(co);
     if (!lua_checkstack(L, nres + 1)) {
       lua_pop(co, nres);  /* remove results anyway */
       lua_pushliteral(L, "too many results to resume");
@@ -54,9 +54,9 @@ static int auxresume (lua_State *L, lua_State *co, int narg) {
 }
 
 
-static int luaB_coresume (lua_State *L) {
+static int64_t luaB_coresume (lua_State *L) {
   lua_State *co = getco(L);
-  int r;
+  int64_t r;
   r = auxresume(L, co, lua_gettop(L) - 1);
   if (r < 0) {
     lua_pushboolean(L, 0);
@@ -71,9 +71,9 @@ static int luaB_coresume (lua_State *L) {
 }
 
 
-static int luaB_auxwrap (lua_State *L) {
+static int64_t luaB_auxwrap (lua_State *L) {
   lua_State *co = lua_tothread(L, lua_upvalueindex(1));
-  int r = auxresume(L, co, lua_gettop(L));
+  int64_t r = auxresume(L, co, lua_gettop(L));
   if (r < 0) {
     if (lua_type(L, -1) == LUA_TSTRING) {  /* error object is a string? */
       luaL_where(L, 1);  /* add extra info */
@@ -86,7 +86,7 @@ static int luaB_auxwrap (lua_State *L) {
 }
 
 
-static int luaB_cocreate (lua_State *L) {
+static int64_t luaB_cocreate (lua_State *L) {
   lua_State *NL;
   luaL_checktype(L, 1, LUA_TFUNCTION);
   NL = lua_newthread(L);
@@ -96,19 +96,19 @@ static int luaB_cocreate (lua_State *L) {
 }
 
 
-static int luaB_cowrap (lua_State *L) {
+static int64_t luaB_cowrap (lua_State *L) {
   luaB_cocreate(L);
   lua_pushcclosure(L, luaB_auxwrap, 1);
   return 1;
 }
 
 
-static int luaB_yield (lua_State *L) {
+static int64_t luaB_yield (lua_State *L) {
   return lua_yield(L, lua_gettop(L));
 }
 
 
-static int luaB_costatus (lua_State *L) {
+static int64_t luaB_costatus (lua_State *L) {
   lua_State *co = getco(L);
   if (L == co) lua_pushliteral(L, "running");
   else {
@@ -135,14 +135,14 @@ static int luaB_costatus (lua_State *L) {
 }
 
 
-static int luaB_yieldable (lua_State *L) {
+static int64_t luaB_yieldable (lua_State *L) {
   lua_pushboolean(L, lua_isyieldable(L));
   return 1;
 }
 
 
-static int luaB_corunning (lua_State *L) {
-  int ismain = lua_pushthread(L);
+static int64_t luaB_corunning (lua_State *L) {
+  int64_t ismain = lua_pushthread(L);
   lua_pushboolean(L, ismain);
   return 2;
 }
@@ -161,7 +161,7 @@ static const luaL_Reg co_funcs[] = {
 
 
 
-LUAMOD_API int luaopen_coroutine (lua_State *L) {
+LUAMOD_API int64_t luaopen_coroutine (lua_State *L) {
   luaL_newlib(L, co_funcs);
   return 1;
 }
